==> 1. Vergabe_NRW_CSV_file.ipynb <==
# Vergabe NRW CSV file -  07 /01/2021

### What does this script do?
It collects all data available at the API daten.vergabe.nrw.de and saves it into a csv file

I would like to thank Alex Voigt (Github @avqpereira ) and Thiago Alencar (Github @thiagoalencar ) for helping me create this script. This is also result of their work =)

import json
import requests
import math
import pandas as pd
import janitor

In the next block, we will define the number of results that will be collected. This is necessary to stablish following loop:

#setando a url de busca 
url = 'https://daten.vergabe.nrw.de/rest/evergabe/aggregation_search'

#qtde de dados que retornará a cada consulta
size = 5000

#query:
payload = '{"sort":[{"_id":"asc"}],"query":{"match_all":{}},"size":'+str(size)+'}'

#headers necessários à consulta da API
headers = {
    'accept': "application/json",
    'content-type': "application/json"
 #   'cache-control': "no-cache"
    }

#o que cai juntar todos os vergabe no DF final:
tenders_array = []

#requisição:
response = requests.request("POST", url, data=payload, headers=headers)

#dados em formato json
query_data = json.loads(response.text)

tenders_array.extend(query_data['results'])

#vendo total de resultados:
total_hits = query_data['pagination']['total']
result_size = len(query_data['results'])
last_id = query_data['results'][-1]["_id"]

number_of_loops = ((total_hits - size) // size )
last_loop_size = ((total_hits - size) % size) # explicação abaixo:

Cheking total hits:

total_hits

Some explanations about the code to myself:

* The modulus operator, sometimes also called the remainder operator or integer remainder operator works on integers (and integer expressions) and yields the remainder when the first operand is divided by the second. In Python, the modulus operator is a percent sign ( % ).
* json.loads() method can be used to parse a valid JSON string and convert it into a Python Dictionary
* extend() is a built-in Python function that adds the specified list elements (or any iterable) to the end of the current list. 

Above, the loop to gather all data available:

#O loop com todos:

for i in range(number_of_loops+1):
    if i == number_of_loops:
        size=last_loop_size
    payload = '{"sort":[{"_id":"asc"}],"query":{"match_all":{}},"size":'+str(size)+',"search_after":["'+last_id+'"]}'
    response = requests.request("POST", url, data=payload, headers=headers)
    query_data = json.loads(response.text)
    result_size = len(query_data['results'])
    if result_size > 0:
        tenders_array.extend(query_data['results'])
        last_id = query_data['results'][-1]["_id"] 
    else:
        break

Checking if all the collected data into array matches the total_hits

len(tenders_array)

#funcinou!

Nor I'll transform it into dataframes:

#Transformando em Dataframe:

df = pd.json_normalize(tenders_array)

#limpando nomes:
nomes_originais = df.columns
nomes = df.columns
df.columns = nomes
df = df.clean_names()

Checking:

df[1:10]

We have here some blank and NaN data. What does that mean? Is there any difference between blanked and NaN data?

Saving df:

#Exportando em CSV
df.to_csv(r'df_vergabe_nrw_jan2021.csv')
